+++ agp-compatibility/agp-compatibility-7-2/src/main/java/co/elastic/apm/android/agp/Agp72CompatibilityManager.java
@@ -3,7 +3,7 @@ package co.elastic.apm.android.agp;
 import org.gradle.api.Project;
 
 import co.elastic.apm.android.agp.api.AgpCompatibilityManager;
-import co.elastic.apm.android.agp.api.AgpUseCase;
+import co.elastic.apm.android.agp.api.usecase.AgpUseCase;
 
 public class Agp72CompatibilityManager extends AgpCompatibilityManager {
