+++ agp-compatibility/agp-compatibility-api/src/main/java/co/elastic/apm/android/agp/api/AgpCompatibilityManager.java
@@ -2,6 +2,8 @@ package co.elastic.apm.android.agp.api;
 
 import org.gradle.api.Project;
 
+import co.elastic.apm.android.agp.api.usecase.AgpUseCase;
+
 public abstract class AgpCompatibilityManager {
     protected final Project project;
